cleanup: move lightning client to seperate package
